Factor out URLFile
